pyproject.toml
[tool.poetry]
name = "virtual-market-dev"
version = "0.1.0"
description = "this is a virtual market bot."
authors = ["hawk-tomy <hawk201000211101@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
